POFIM-5 Tydeliggjort om dto er respone eller request, og definert alle records i  InntektsmeldingDialogRest klassen.
