--- src/main/java/no/nav/familie/inntektsmelding/imdialog/rest/OrganisasjonInfoDto.java
@@ -1,6 +1,0 @@
-package no.nav.familie.inntektsmelding.imdialog.rest;
-
-import jakarta.validation.constraints.NotNull;
-
-public record OrganisasjonInfoDto (@NotNull String organisasjonNavn, @NotNull String organisasjonNummer){
-}
