--- src/main/java/no/nav/familie/inntektsmelding/imdialog/rest/PersonInfoDto.java
@@ -1,5 +1,0 @@
-package no.nav.familie.inntektsmelding.imdialog.rest;
-
-import jakarta.validation.constraints.NotNull;
-
-public record PersonInfoDto(@NotNull String navn, @NotNull String fødselsnummer, @NotNull String aktørId) {}
